Performing connected components is now fast because it returns a generator and we used to sort it, but because of our upper bound for GCC size, we don't need to sort it so the performance hit shows up in the next step of slash burn
