+++ STRUCTURE_DISCOVERY_PY/profiling.py
@@ -1,0 +1,15 @@
+import cProfile
+
+
+def profiler(func):
+    def profiled_func(*args, **kwargs):
+        profile = cProfile.Profile()
+        try:
+            profile.enable()
+            result = func(*args, **kwargs)
+            profile.disable()
+            return result
+        finally:
+            profile.print_stats()
+    return profiled_func
+
