+++ STRUCTURE_DISCOVERY_PY/structures.py
@@ -1,7 +1,7 @@
-from operator import itemgetter
 from math import log
 import numpy as np
 import networkx as nx
+from operator import itemgetter
 
 
 def ln(n):
